Devhelp: add tests for convert_devhelp_to_qch()
