+++ tests/devhelp_qch_data/file-list.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?><files>
+<file>./en/path1.html</file>
+<file>./en/path2.html</file>
+<file>./en/path3.html</file>
+<file>./en/path4.html</file>
+<file>./en/path5.html</file>
+<file>./en/path6.html</file>
+<file>./en/path7.html</file>
+</files>
